Move pad_tx from txtools.py to blocktools.py
Since that function was the whole content of txtools.py
we removed the file completely.
