+++ qa/rpc-tests/electrum_scripthash_gethistory.py
@@ -9,8 +9,7 @@ from test_framework.loginit import logging
 from test_framework.electrumutil import *
 from test_framework.nodemessages import COIN
 from test_framework.blocktools import create_coinbase, create_block, \
-    create_transaction
-from test_framework.txtools import pad_tx
+    create_transaction, pad_tx
 from test_framework.mininode import (
     P2PDataStore,
     NodeConn,
